feat: directory clean up
