+++ mobile/core/data/dao/.gitignore
