+++ mobile/core/data/dao/proguard-rules.pro
