+++ mobile/core/data/dao/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/outside/datasource/SampleDataSource.kt
@@ -1,0 +1,10 @@
+package co.kr.parkjonghun.whatishedoingwithandroid.outside.datasource
+
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.model.S1
+
+// TODO: using flow to follow the UDF.
+class SampleDataSource {
+    fun getS1(): S1 {
+        return S1(s = "sample data.")
+    }
+}
