+++ mobile/core/data/dao/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/outside/datastore/DataStores.kt
@@ -1,4 +1,4 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.data.model
+package co.kr.parkjonghun.whatishedoingwithandroid.outside.datastore
 
 data object DataStores {
     const val Sample = "sample"
