+++ mobile/core/data/dao/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/outside/datastore/Shortcuts.kt
@@ -1,10 +1,9 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.data.extension
+package co.kr.parkjonghun.whatishedoingwithandroid.outside.datastore
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
-import co.kr.parkjonghun.whatishedoingwithandroid.data.model.DataStores
 
 val Context.sampleDataStore: DataStore<Preferences> by preferencesDataStore(name = DataStores.Sample)
 val Context.userDataStore: DataStore<Preferences> by preferencesDataStore(name = DataStores.User)
