+++ mobile/core/data/dao/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/outside/di/DataSourceModule.kt
@@ -1,0 +1,23 @@
+package co.kr.parkjonghun.whatishedoingwithandroid.outside.di
+
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.datasource.PreferencesDataSource
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.datasource.PreferencesDataSourceImpl
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.datasource.RemoteDataSource
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.datasource.RemoteDataSourceImpl
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.datasource.SampleDataSource
+import org.koin.android.ext.koin.androidContext
+import org.koin.dsl.module
+
+val dataSourceModule = module {
+    single<PreferencesDataSource> {
+        PreferencesDataSourceImpl(
+            context = androidContext(),
+        )
+    }
+
+    single<RemoteDataSource> {
+        RemoteDataSourceImpl()
+    }
+
+    single { SampleDataSource() }
+}
