+++ mobile/core/data/dao/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/outside/model/S1.kt
@@ -1,0 +1,3 @@
+package co.kr.parkjonghun.whatishedoingwithandroid.outside.model
+
+data class S1(val s: String)
