+++ mobile/core/data/interior/build.gradle.kts
@@ -9,5 +9,5 @@ android.namespace = "co.kr.parkjonghun.whatishedoingwithandroid.data.interior"
 
 dependencies {
     implementation(project(":core:domain:service"))
-    implementation(project(":core:data:surface"))
+    implementation(project(":core:data:dao"))
 }
