+++ mobile/core/data/interior/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/inside/mapper/SampleRepositoryMapper.kt
@@ -1,6 +1,6 @@
 package co.kr.parkjonghun.whatishedoingwithandroid.inside.mapper
 
-import co.kr.parkjonghun.whatishedoingwithandroid.data.model.S1
+import co.kr.parkjonghun.whatishedoingwithandroid.outside.model.S1
 import co.kr.parkjonghun.whatishedoingwithandroid.service.model.sample.Sample
 
 fun S1.toDomainModel(): Sample = Sample(stuff = s)
