--- mobile/core/data/surface/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/data/gateway/datasource/SampleDataSource.kt
@@ -1,10 +1,0 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.data.gateway.datasource
-
-import co.kr.parkjonghun.whatishedoingwithandroid.data.model.S1
-
-// TODO: using flow to follow the UDF.
-class SampleDataSource {
-    fun getS1(): S1 {
-        return S1(s = "sample data.")
-    }
-}
