--- mobile/core/data/surface/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/data/model/S1.kt
@@ -1,3 +1,0 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.data.model
-
-data class S1(val s: String)
